+++ sunray/src/skytraq/skytraq.cpp
@@ -3,9 +3,9 @@
 // Licensed GPLv3 for open source use
 // or Grau GmbH Commercial License for commercial use (http://grauonline.de/cms2/?page_id=153)
 
+#include "../../config.h"
 #include "Arduino.h"
 #include "skytraq.h"
-#include "../../config.h"
 
 
 //#define GPS_DUMP 1
@@ -128,7 +128,7 @@ void SKYTRAQ::parseBinary(int b)
   else if (this->state == GOT_ID) {
 
       this->addchk(b);
-      if (this->count < sizeof(this->payload)){
+      if (this->count < (int)sizeof(this->payload)){
         this->payload[this->count] = b; 
       } 
       this->count += 1;
